+++ example/lib/main.dart
@@ -28,9 +28,8 @@ class _HomePageState extends State<HomePage> {
       body: ListView(
         children: <Widget>[
           StylableCalendar(
-            specialDays: [20, 13, 6, 10],
-            highlightedDays: [1, 4, 5],
-            isLoading: false,
+//            specialDays: [20, 13, 6, 10],
+//            highlightedDays: [1, 4, 5],
 //            primaryColor: Theme.of(context).primaryColor,
 //            primaryColorDark: Theme.of(context).primaryColorDark,
 //            secondaryColor: Theme.of(context).primaryColorLight,